move getexportvalue into util
